split up huge project index spec into separate files
